# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriTRUST Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /agrihaul/shipments:
    get:
      operationId: listShipments
      tags: [agrihaul]
      summary: List shipments
      responses:
        "200":
          description: Shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [agrihaul]
      summary: Book a shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Shipment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /agrihaul/shipments/{id}:
    get:
      operationId: getShipment
      tags: [agrihaul]
      summary: Get shipment details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  # ── Community ─────────────────────────────────────────────────────────────
...
  /accounting/transactions:
    get:
      operationId: listTransactions
      tags: [accounting]
      summary: List farm transactions
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string }
        - name: to
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [accounting]
      summary: Record a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List theft alerts
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TheftAlert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Report theft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TheftAlertInput"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftAlert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get theft alert
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftAlert"

  # ── Disputes ──────────────────────────────────────────────────────────────
...
  /academy/courses:
    get:
      operationId: listCourses
      tags: [academy]
      summary: List academy courses
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: level
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /academy/courses/{id}:
    get:
      operationId: getCourse
      tags: [academy]
      summary: Get course detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /academy/courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [academy]
      summary: Enroll in a course
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [farmer, buyer, investor, service_provider] }
        region: { type: string }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        region: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isVerified: { type: boolean }
        farmScoreRating: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
...
    Shipment:
      type: object
      required: [id, origin, destination, cargoType, weightKg, status, userId, createdAt]
      properties:
        id: { type: integer }
        origin: { type: string }
        destination: { type: string }
        cargoType: { type: string }
        weightKg: { type: number }
        status: { type: string }
        trackingCode: { type: ["string", "null"] }
        estimatedDelivery: { type: ["string", "null"] }
        actualDelivery: { type: ["string", "null"] }
        driverName: { type: ["string", "null"] }
        driverPhone: { type: ["string", "null"] }
        costNAD: { type: ["number", "null"] }
        userId: { type: integer }
        createdAt: { type: string, format: date-time }
    ShipmentInput:
      type: object
      required: [origin, destination, cargoType, weightKg]
      properties:
        origin: { type: string }
        destination: { type: string }
        cargoType: { type: string }
        weightKg: { type: number }
        notes: { type: string }

    # ── Community Schemas ─────────────────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, type, amount, category, description, date, userId]
      properties:
        id: { type: integer }
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        category: { type: string }
        description: { type: string }
        date: { type: string, format: date }
        referenceId: { type: ["string", "null"] }
        userId: { type: integer }
        createdAt: { type: string, format: date-time }
    TransactionInput:
      type: object
      required: [type, amount, category, description, date]
      properties:
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        category: { type: string }
        description: { type: string }
        date: { type: string, format: date }
        referenceId: { type: string }
...
    TheftAlert:
      type: object
      required: [id, title, description, region, itemType, status, reporterId, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        region: { type: string }
        itemType: { type: string }
        status: { type: string }
        imageUrl: { type: ["string", "null"] }
        rewardOffered: { type: ["number", "null"] }
        reporterId: { type: integer }
        reporterName: { type: ["string", "null"] }
        resolvedAt: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    TheftAlertInput:
      type: object
      required: [title, description, region, itemType]
      properties:
        title: { type: string }
        description: { type: string }
        region: { type: string }
        itemType: { type: string }
        imageUrl: { type: string }
        rewardOffered: { type: number }

    # ── Dispute Schemas ───────────────────────────────────────────────────────
...
    Course:
      type: object
      required: [id, title, category, level, durationHours, enrollmentCount, rating, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        level: { type: string }
        durationHours: { type: number }
        thumbnailUrl: { type: ["string", "null"] }
        instructorName: { type: string }
        enrollmentCount: { type: integer }
        rating: { type: number }
        isFree: { type: boolean }
        modules:
          type: array
          items:
            type: object
            required: [title, lessonCount]
            properties:
              title: { type: string }
              lessonCount: { type: integer }
        createdAt: { type: string, format: date-time }
    Enrollment:
      type: object
      required: [id, courseId, userId, status, progressPercent, enrolledAt]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        courseTitle: { type: ["string", "null"] }
        userId: { type: integer }
        status: { type: string }
        progressPercent: { type: number }
        completedAt: { type: ["string", "null"] }
        certificateUrl: { type: ["string", "null"] }
        enrolledAt: { type: string, format: date-time }